+++ src/App.tsx
@@ -1,11 +1,11 @@
 import './App.css';
-import Practice5 from './components/templates/Practice5';
+import Practice4 from './components/templates/Practice4';
 
 function App() {
   return (
     <>
       <header className="bg-[#94A3B8] text-center p-[20px] text-4xl text-[#F9FAFB]">React-v3</header>
-      <Practice5 />
+      <Practice4 />
     </>
   );
 }
